Move arithmatic routines to their own module.
